Add device id annotation to the allocate response
Added aarna.com/sriov_device_id annotation to the device plugin
container response for the amcop to be able to identify the sriov
vf allocated to the pod. This is needed since amcop does not have
any other way to identify the vf so that it could plug its rep
in the bluefield. this flow must be outside the cni flow, since
the cni would fail with kata containers.